requirements.txt
anyio==4.4.0
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==43.0.0
duckduckgo_search==6.2.4
faiss-cpu==1.8.0.post1
filelock==3.15.4
Flask==3.0.3
fsspec==2024.6.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.24.3
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
lxml==5.2.2
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
ollama==0.3.0
packaging==24.1
pdfminer.six==20231228 
pdfplumber==0.11.2
pillow==10.4.0
primp==0.5.4
pycparser==2.22
PyPDF2==3.0.1
pypdfium2==4.30.0
python-docx==1.1.2
python-dotenv==1.0.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
safetensors==0.4.3
scikit-learn==1.5.1
scipy==1.14.0
sentence-transformers==3.0.1
setuptools==72.1.0
sniffio==1.3.1
soupsieve==2.5
sympy==1.13.1
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.3.0
torchaudio==2.3.0
torchvision==0.18.0
tqdm==4.66.4
transformers==4.43.3
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.3

